觉得不太合适可以取消了，弄的时候报了很多错

+++ Backend/Controllers/XboxController.cs
@@ -268,7 +268,7 @@ public class XboxController : ControllerBase
             {
                 userPlatformBinding = new UserPlatformBinding
                 {
-                    UserId = userId,
+                    UserId = (int)userId,
                     PlatformId = XBOX_PLATFORM_ID,
                     PlatformUserId = xboxUser.Xuid,
                     BindingStatus = true,
